refactor: Rename Storage and FileHandler class for better clarity
Storage was originally a wrapper around an ArrayList of tasks, which
is more accurately represented as TaskList.

FileHandler, responsible for saving/loading tasks, has been renamed
to Storage to better reflect its role.
