+++ src/main/java/oogway/storage/TaskList.java
@@ -1,0 +1,33 @@
+package oogway.storage;
+
+import oogway.tasks.Task;
+
+import java.util.ArrayList;
+
+public class TaskList {
+    private final ArrayList<Task> tasks = new ArrayList<>();
+
+    public void addTask(Task task) {
+        tasks.add(task);
+    }
+
+    public Task getTask(int index) {
+        return tasks.get(index);
+    }
+
+    public Task deleteTask(int index) {
+        return tasks.remove(index);
+    }
+
+    public int getTaskCount() {
+        return tasks.size();
+    }
+
+    public boolean isEmpty() {
+        return tasks.isEmpty();
+    }
+
+    public ArrayList<Task> getTasks() {
+        return new ArrayList<>(tasks);
+    }
+}
